Update: richMenu switching function
