+++ functions/src/triggers/https/line-bot/handlers/index.ts
@@ -3,12 +3,13 @@ import { lineClient } from '~/utils/line'
 import { errorLogger } from '~/utils/util'
 import { msgOther } from '~line/notice-messages/other'
 import { messageImageHandler } from './image'
+import { messageTextHandler } from './messages'
 
 export const messagesHandler = async (event: MessageEvent): Promise<void> => {
   try {
     switch (event.message.type) {
-      // case 'text':
-      //   return await messageTextHandler(event)
+      case 'text':
+        return await messageTextHandler(event)
       case 'image':
         return await messageImageHandler(event)
       default:
