+++ functions/src/utils/secrets.ts
@@ -4,3 +4,4 @@ import 'dotenv/config'
 
 export const LINE_MESSAGING_CHANNEL_ACCESS_TOKEN = process.env.LINE_MESSAGING_CHANNEL_ACCESS_TOKEN as string
 export const LINE_MESSAGING_CHANNEL_SECRET = process.env.LINE_MESSAGING_CHANNEL_SECRET as string
+export const MY_LINE_USER_ID = process.env.MY_LINE_USER_ID as string
